read the body of response in a separate threadpool (yuk)
right now we have cases where broken http response can cause the
reading of the body to hang forever. the http client expects us to
close the inputstream backing the body in such conditions.

to quote the javadoc:.

"To ensure that all resources associated with the corresponding
exchange are properly released the caller must ensure to either read
all bytes until EOF is reached, or call InputStream.close() if it is
unable or unwilling to do so. Calling close before exhausting the
stream may cause the underlying HTTP connection to be closed and
prevent it from being reused for subsequent operations."